network: add support to interoperate with some network discovery functions
This new module permits to configure DSR wat IP level and the
daemon will discover the MAC addresses of the backends using
libmnl to read from the neighbor table or sending pings to force
the insertion in the table if doesn't exist. It also gets the
local MAC addresses, and even register to the neighbor table if
there is any new event.

Signed-off-by: Laura Garcia <laura.garcia@zevenet.com>