backends: include a new backend state config_error
The service needs different backend information according to
the farm mode, specially for DSR, where some data is needed
to be gathered automatically.

For that reason, we insert a new state to identify if the
backend has some missing data before to be configured in the
virtual service.

This new state is called config_error, and the backend won't
be configured until this state is becoming up again.

Signed-off-by: Laura Garcia <laura.garcia@zevenet.com>