server: add struct nftlb_http_state
This structure conveys the state of the HTTP request handler. This
includes the response code, so we can just return -1 on error from
get_request() and 0 in case of success.

I have renamed 'content' to 'body', 'action' to method and 'response' to
'status_code' to refer to more orthodox http lingo.

As a result, we do not need HTTP_MIN_CONTINUE to return errors from
functions.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>