src: add support of log prefix
The log prefix for several fields are required to customize the
messages in the log files.

This patch enables the following options.

- For farms:
   log-prefix : virtual service connection logging
   new-rtlimit-log-prefix : virtual service new rate limit reached logging
   rst-rtlimit-log-prefix : virtual service reset rate limit reached logging
   est-connlimit-log-prefix : virtual service established connections limit reached logging

- For backends:
   est-connlimit-log-prefix : backends established connections limit reached logging

- For policies:
   log-prefix : policy matching logging

In order to be flexible, several tokens are created:

KNAME : name of the field
TYPE : type of the field
FNAME : farm name
BNAME : backend name (only for backends)
PNAME : policy name (only for policies)

Signed-off-by: nevola <laura.garcia@zevenet.com>