farms: fix rulerize everything stops after wont rulerize
When we found a farm that is not able to be rulerized, the rest
of the farms should continue with the process.

Before this patch, we found:

 list_for_each_entry_safe(f, next, farms, list)
 	ret = ret || farm_rulerize(f);

With this, we're implicitly avoiding to continue with the
rulerization of farms once one of them won't be rulerized,
as _ret_ becomes true and the second condition are not being
processed.

The same issue occurs with the policies rulerization.

Signed-off-by: nevola <laura.garcia@zevenet.com>