farms: support of stateless dnat direct clients
Direct clients connecting to the stateless dnat load balancer
needs to maintain their source MAC address in order to compile
the packet back correctly. In order to solve this, a map will
be used to maintain during a complete request-response the
MAC address of the client.

In addition, this patch adds a new ttl parameter in order to
be able to configure the timeout of this map association.

Signed-off-by: nevola <laura.garcia@zevenet.com>