nft: add option to serialize nft commands
nft commands generated by nftlb are batched in the form of
concatenations and then executed at once in a single
transactions.

This way of execution can produce some errors due to
deletion and creation of elements that requires a kernel
update in some cases.

This patch adds a new option to nftlb in order to serialize
the nft commands instead of batching. This can produce a
little bit of slowness executing the rules but safer against
buggy kernels.

Signed-off-by: nevola <laura.garcia@zevenet.com>