farms: fix stateless dnat source MAC in order to ensure a consistent traffic
During a stateless DNAT, the source address MAC during the outbound traffic
were not set despite this is needed when the output interface is not the
same interface than the input traffic. That behaviour can produce traffic
inconsistencies as the source MAC address is not changed by the output
interface.

This patch adds the MAC source address and ensures the correct traffic
natting to the backends.

Signed-off-by: nevola <laura.garcia@zevenet.com>