sessions: fix sigsegv in session_backend_action() when there is no backend
In cases where the backend is not identified, ex:

	map persist-console-ocp-adm01-arag-com-8443-lx {
		type ipv4_addr : mark
		size 65535
		timeout 1m
		elements = { 10.252.18.204 expires 22s448ms : 0x00000000 }
	}

cause there are no backend available and the session it's not properly
marked, the function session_backend_action() doesn't check properly that
the backend is identified in the session.

Also, the rule to update the session avoids that the connections with empty
mark session of being registered.

Signed-off-by: nevola <laura.garcia@zevenet.com>