elements: improve performance saving and loading high number of elements
With a number of elements around 80k for a blacklist we got 15 minutes
to response a GET request. With this patch we get 6 seconds for the same
case.

I has been discovered that the high penalty of performance is made by
the snprintf() call, so we created in the new tools file a replacement
for such function.

Signed-off-by: nevola <laura.garcia@zevenet.com>