elements: implement performance changes for loading very long blacklists
Loading very long blacklists will result in performance penalty mainly
due to the function element_lookup_by_name() that performs strcmp() to
every element in the list before creating the element. This is not
longer needed as nft will take care of inserting the element if it
doesn't exist, merge in case of overlapping ranges, etc.

Signed-off-by: nevola <laura.garcia@zevenet.com>