tools: introduce support for log output configuration
The log output can be configured to be used syslog, stdout and stderr
or even combinations of syslog with any of the other outputs. In
order to configure this, a new parameter is used during the launch
time of the nftlb daemon with -L <config>.

This will permit to dump the log messages in machines without syslog
installed.

Signed-off-by: nevola <laura.garcia@zevenet.com>