nft: nft command error reloading forward chain
Activating the forward logs we can face the following error:

"Error: Could not process rule: Device or resource busy" while trying
to set again an element in the forward service map.

The problem comes from the 'nft_chain' flag was not activating the
forward chain flag properly and for this the function
run_farm_rules_gen_vsrv() will be used.

Signed-off-by: nevola <laura.garcia@zevenet.com>