addresses: add support of policies in addresses
Policies can be assigned to addresses, so no need to be assigned
to a farm. A request to /addresses can be done with a configuration
like this one:

{
	"addresses" : [
		{
			"name" : "lonelyaddr",
			"family" : "ipv4",
			"ip-addr" : "127.0.0.1",
			"ports" : "777",
			"protocol" : "udp",
			"policies" : [
				{
					"name" : "black001"
				}
			]
		}
	]
}

Note that the ip address could be configured in the system or not, but
in this last case, the inbound interface need to be specified.

Signed-off-by: nevola <laura.garcia@zevenet.com>