src: add support of configurable verdicts for limits and security lists for farms and addresses
With this new option, log, drop or accept are configurable actions
when a matching is performed. Combinations of them are allowed.

"verdict" : "log" - only logging is performed

"verdict" : "log drop" - logging and then drop for limits and
blacklists, only logging for whitelists

"verdict" : "log drop accept" - logging and drop or accept for
all limits or security lists.

Signed-off-by: nevola <laura.garcia@zevenet.com>