nft: fix whitelist prioritization to switch to drop policy
With this patch, when used whitelists no blacklists are configured
in the virtual service and a final drop rule is set in order to
switch to a drop policy by default. When no whitelists are configured,
then the accept policy will be set again by default.

Also, fix return 0 to continue for the case of other policies should
be configured for whitelists or blacklists.

Signed-off-by: nevola <laura.garcia@zevenet.com>