src: avoid resource busy when using different farm structures
Backends map were not handled properly when using different farm
structures in the service maps. Counters also need to be updated
properly in order to know when it should be deleted. Also,
priorities functions need to be adjusted in order to delete
the backends properly and to be updated in the counter.

Signed-off-by: nevola <laura.garcia@zevenet.com>