nft: apply sequential order for policies
Delete priority of lists and apply policies in sequential order of
addition, independently of being white or black lists. In addition,
default verdict is not applied, the admin should create a dummy list
with the element "0.0.0.0" as whitelist in order to apply a default
accept or a blacklist in order to apply a drop.

Signed-off-by: nevola <laura.garcia@zevenet.com>