nft: do not insert sessions entries if no backend is available and the farm uses masquerade
It is needed to manage the masquerade bit in order to avoid to insert any
entry in the sessions map when there is no backend available. Otherwise,
sessions could be inserted assigned incorrectly to a "masquerade bit" mark.

table ip nftlb {
        map persist-myfarm-144-lx4nat {
                type ipv4_addr : mark
                size 65535
                timeout 1h
                elements = { 192.168.233.1 timeout 1h expires
59m57s532ms : 0x80000000 }
        }
}

Filter the masquerade mark only for farms that uses it.

Signed-off-by: nevola <laura.garcia@zevenet.com>