server: enable _patch_ verb for policy farms
In order to speed up the configuration of farm policies, in the case
of a big amount of policies configured in a given farm, instead of
deleting the policies and apply again the new configuration we
enable the PATCH verb in the API server.

With the PATCH /farms/<my_farm>/policies/ call, the service will
take care of stopping all the associated nft rules, delete the
objects and apply the new ones. The JSON structure it is required
to apply to the farm with the policies list.

{
        "farms": [
                {
                        "name": "newfarm",
                        "policies": [
                                {
                                        "name": "aa_white1"
                                },
                                {
                                        "name": "aa_white2"
                                }
                        ]
                }
        ]
}

Signed-off-by: nevola <laura.garcia@zevenet.com>